* hotfix/1.47.1:
  Remove unused type
  #19 Pass passport as first middlwware to loging route
  #19 Refactor catch clouses, and refactor vars for better reading

+++ routes/controllers/todos.ts
@@ -3,6 +3,7 @@ import express from 'express';
 import {getTodos, getTodo, createTodo, patchTodo, deleteTodo}  from '../middleware/validateTodo';
 import validate from '../middleware/validate';
 import prisma from '../../db/prisma'
+import HttpErrors from 'http-errors'
 const router = express.Router();
 
 // READ
@@ -10,12 +11,14 @@ router.get('/', getTodos(), validate, async (req: express.Request, res: express.
   if (!req.body.user_id) console.error('quote ID is required');
 
   try {
-    const data = await prisma.todo.findMany({where: { user_id: req.body.user_id}});
+    const data = await prisma.todo.findMany({where: {user_id: req.body.user_id}});
     return res.json({errors: false, data: data});
-  } catch (err: any) {
-    return res.status(500).json({errors: [err.message], data: {}});
+  } catch (error) {
+    if (error instanceof Error) return res.json({errors: [error.message], data: {}});
+    return res.json(error)
   }
 });
+
 // CREATE
 router.post('/', createTodo(), validate, async (req: express.Request, res: express.Response) => {
   try {
@@ -28,59 +31,71 @@ router.post('/', createTodo(), validate, async (req: express.Request, res: expre
       },
     });
     return res.status(201).json({errors: false, data: result});
-  } catch (err: any) {
-    return res.status(500).json({errors: [err.message], data: {}});
+  } catch (error) {
+    if (error instanceof Error) return res.json({errors: [error.message], data: {}});
+    return res.json(error)
   }
 });
+
 // READ
 router.get('/:id([0-9]+)', getTodo(), validate, async (req: express.Request, res: express.Response) => {
   if (!req.params.id) console.error('quote ID is required');
 
   try {
-    const data = await prisma.todo.findMany({where: { id: Number(req.params.id)}});
-    if (!data) {
-      return res.status(404).json({errors: true, data: {}});
-    } else {
-      return res.json({errors: false, data: data});
-    }
-  } catch (error: any) {
-    return res.status(500).json({errors: [error.message], data: {}});
+    const data = await prisma.todo.findMany({where: {id: Number(req.params.id)}});
+
+    // no match by id
+    if (!data) throw new HttpErrors.NotFound('Record not found')
+
+    return res.json({errors: false, data: data});
+  } catch (error) {
+    if (error instanceof Error) return res.json({errors: [error.message], data: {}});
+    return res.json(error)
   }
 });
+
 // UPDATE
 router.patch('/:id([0-9]+)', patchTodo(), validate, async (req: express.Request, res: express.Response) => {
   if (!req.params.id) console.error('todo ID is required');
+  const { title, completed, user_id } = req.body
 
   try {
     const data = await prisma.todo.findUnique({where: {id: Number(req.params.id)}});
-    if (data) {
-      const { title, completed, user_id } = req.body
-      const toUpdate = Object.assign(req.body, {updated_at: new Date().toISOString()});
-      await prisma.todo.update({
-        where: {id: Number(data.id)},
-        data: {
-          title,
-          completed,
-          user_id
-        }
+
+    // no match by id
+    if (!data) throw new HttpErrors.NotFound('Record not found')
+
+    // const toUpdate = Object.assign(req.body, {updated_at: new Date().toISOString()});
+    const updated = await prisma.todo.update({
+      where: {id: Number(data.id)},
+      data: {
+        title,
+        completed,
+        user_id
+      }
     });
-      // await new Todo({id: data.id}).save(toUpdate, {patch: true});
-      return res.json({ errors: false, data: data, message: 'Todo updated'});
-    } else {
-      return res.status(404).json({errors: 'Todo not found', data: {}});
-    }
-  } catch (err: any) {
-    return res.status(500).json({ errors: [err.message] });
+
+    return res.json({ errors: false, data: updated, message: 'Todo updated'});
+  } catch (error) {
+    if (error instanceof Error) return res.json({errors: [error.message], data: {}});
+    return res.json(error)
   }
 });
+
 // DELETE
 router.delete('/:id([0-9]+)', deleteTodo(), validate, async (req: express.Request, res: express.Response) => {
   if (!req.params.id) console.error('Todo ID is required');
   try {
-    const data = await prisma.todo.delete({ where: {id: Number(req.params.id)}});
+    const item = await prisma.todo.findUnique({where: {id: Number(req.params.id)}});
+
+    // no match by id
+    if (!item) throw new HttpErrors.NotFound('Record not found')
+
+    const data = await prisma.todo.delete({where: {id: Number(req.params.id)}});
     return res.json({errors: false, data: data, message: `Todo removed, id: ${req.params.id}`});
-  } catch (err: any) {
-    return res.status(500).json({errors: [err.message]});
+  } catch (error) {
+    if (error instanceof Error) return res.json({errors: [error.message], data: {}});
+    return res.json(error)
   }
 });
